backend/requirements.txt
fastapi
uvicorn
sqlmodel
werkzeug
python-multipart
requests
pydantic
google-genai